+++ tests/helpers.js
@@ -1,6 +1,6 @@
 // eslint-disable-next-line import/prefer-default-export
-export async function sleep() {
+export async function sleep(ms = 0) {
   return new Promise((resolve) => {
-    setTimeout(resolve, 0);
+    setTimeout(resolve, ms);
   });
 }